rename and move PlatformWiring
