--- platform-sdk/swirlds-platform-core/src/main/java/com/swirlds/platform/builder/PlatformBuildingBlocks.java
@@ -1,122 +1,0 @@
-// SPDX-License-Identifier: Apache-2.0
-package com.swirlds.platform.builder;
-
-import static java.util.Objects.requireNonNull;
-
-import com.hedera.hapi.node.base.SemanticVersion;
-import com.swirlds.common.context.PlatformContext;
-import com.swirlds.common.notification.NotificationEngine;
-import com.swirlds.component.framework.model.WiringModel;
-import com.swirlds.platform.state.ConsensusStateEventHandler;
-import com.swirlds.platform.wiring.PlatformComponents;
-import com.swirlds.platform.wiring.PlatformCoordinator;
-import com.swirlds.state.StateLifecycleManager;
-import com.swirlds.state.merkle.VirtualMapState;
-import com.swirlds.virtualmap.VirtualMap;
-import edu.umd.cs.findbugs.annotations.NonNull;
-import java.security.SecureRandom;
-import java.util.concurrent.atomic.AtomicReference;
-import java.util.function.Supplier;
-import org.hiero.base.concurrent.BlockingResourceProvider;
-import org.hiero.consensus.event.IntakeEventCounter;
-import org.hiero.consensus.gossip.ReservedSignedStateResult;
-import org.hiero.consensus.hashgraph.FreezePeriodChecker;
-import org.hiero.consensus.model.node.KeysAndCerts;
-import org.hiero.consensus.model.node.NodeId;
-import org.hiero.consensus.monitoring.FallenBehindMonitor;
-import org.hiero.consensus.roster.RosterHistory;
-import org.hiero.consensus.state.SavedStateController;
-import org.hiero.consensus.state.nexus.SignedStateNexus;
-import org.hiero.consensus.state.signed.ReservedSignedState;
-import org.hiero.consensus.status.StatusActionSubmitter;
-
-/**
- * This record contains core utilities and basic objects needed to build a platform. It should not contain any platform
- * components.
- *
- * @param platformComponents                     the wiring for this platform
- * @param platformContext                        the context for this platform
- * @param model                                  the wiring model for this platform
- * @param keysAndCerts                           an object holding all the public/private key pairs and the CSPRNG state
- *                                               for this member
- * @param selfId                                 the ID for this node
- * @param mainClassName                          the name of the app class inheriting from SwirldMain
- * @param swirldName                             the name of the swirld being run
- * @param appVersion                             the current version of the running application
- * @param initialState                           the initial state of the platform
- * @param rosterHistory                          the roster history provided by the application to use at startup
- * @param intakeEventCounter                     counts events that have been received by gossip but not yet inserted
- *                                               into gossip event storage, per peer
- * @param secureRandomSupplier                   a source of secure random number generator instances
- * @param freezeChecker                          a predicate that determines if a timestamp is in the freeze period
- * @param consensusEventStreamName               a part of the name of the directory where the consensus event stream is
- *                                               written
- * @param notificationEngine                     for sending notifications to the application (legacy pattern)
- * @param statusActionSubmitterReference         a reference to the status action submitter, this can be deleted once
- *                                               platform status management is handled by the wiring framework
- * @param stateLifecycleManager                  responsible for the mutable state, this is exposed here due to
- *                                               reconnect
- * @param execution                              the instance of the execution layer, which allows consensus to interact
- *                                               with the execution layer
- * @param fallenBehindMonitor                    an instance of the fallenBehind Monitor which tracks if the node has fallen behind
- * @param reservedSignedStateResultPromise             a shared data structure that Gossip and the ReconnectController will use provide
- *                                               and obtain a reference to a ReservedSignedState
- * @param platformCoordinator                    the platform coordinator, which allows components to trigger platform status changes
- * @param latestImmutableStateNexus              a nexus for accessing the latest immutable state
- * @param transactionOffsetNanos                 nanoseconds to add to the first transaction's timestamp in an event,
- *                                               computed by the execution layer from its configuration
- * @param savedStateController                   the controller for saving states to disk
- */
-public record PlatformBuildingBlocks(
-        @NonNull PlatformComponents platformComponents,
-        @NonNull PlatformContext platformContext,
-        @NonNull WiringModel model,
-        @NonNull KeysAndCerts keysAndCerts,
-        @NonNull NodeId selfId,
-        @NonNull String mainClassName,
-        @NonNull String swirldName,
-        @NonNull SemanticVersion appVersion,
-        @NonNull ReservedSignedState initialState,
-        @NonNull RosterHistory rosterHistory,
-        @NonNull IntakeEventCounter intakeEventCounter,
-        @NonNull Supplier<SecureRandom> secureRandomSupplier,
-        @NonNull FreezePeriodChecker freezeChecker,
-        @NonNull String consensusEventStreamName,
-        @NonNull NotificationEngine notificationEngine,
-        @NonNull AtomicReference<StatusActionSubmitter> statusActionSubmitterReference,
-        @NonNull StateLifecycleManager<VirtualMapState, VirtualMap> stateLifecycleManager,
-        @NonNull ConsensusStateEventHandler consensusStateEventHandler,
-        @NonNull ExecutionLayer execution,
-        @NonNull FallenBehindMonitor fallenBehindMonitor,
-        @NonNull BlockingResourceProvider<ReservedSignedStateResult> reservedSignedStateResultPromise,
-        @NonNull PlatformCoordinator platformCoordinator,
-        @NonNull SignedStateNexus latestImmutableStateNexus,
-        long transactionOffsetNanos,
-        @NonNull SavedStateController savedStateController) {
-    public PlatformBuildingBlocks {
-        requireNonNull(platformComponents);
-        requireNonNull(platformContext);
-        requireNonNull(model);
-        requireNonNull(keysAndCerts);
-        requireNonNull(selfId);
-        requireNonNull(mainClassName);
-        requireNonNull(swirldName);
-        requireNonNull(appVersion);
-        requireNonNull(initialState);
-        requireNonNull(rosterHistory);
-        requireNonNull(intakeEventCounter);
-        requireNonNull(secureRandomSupplier);
-        requireNonNull(freezeChecker);
-        requireNonNull(consensusEventStreamName);
-        requireNonNull(notificationEngine);
-        requireNonNull(statusActionSubmitterReference);
-        requireNonNull(stateLifecycleManager);
-        requireNonNull(consensusStateEventHandler);
-        requireNonNull(execution);
-        requireNonNull(fallenBehindMonitor);
-        requireNonNull(reservedSignedStateResultPromise);
-        requireNonNull(platformCoordinator);
-        requireNonNull(latestImmutableStateNexus);
-        requireNonNull(savedStateController);
-    }
-}
